test(range): migrate color
