+++ core/src/components/range/test/color/range.e2e.ts-snapshots/range-color-ios-ltr-Mobile-Chrome-linux.png
