+++ core/src/components/range/test/color/range.e2e.ts-snapshots/range-color-ios-ltr-Mobile-Safari-linux.png
