+++ core/src/components/range/test/color/range.e2e.ts-snapshots/range-color-md-ltr-Mobile-Chrome-linux.png
